Refactor PokemonObject and FloatingText to use the Objects they extend
This commit fixes the way PokemonObject and FloatingText are implemented
so that it properly makes use of the `Phaser.GameObjects` types they extend.

Before this, they would create separate objects inside the contained scenes,
which then had to manually be cleaned up on `destroy()`. This commit changes
the Objects to utilise the core object they extend (`Sprite` for `PokemonObject`,
and `Text` for `FloatingText`) and set position based off that.

Scenes using the component need to use `add.existing` to add these components,
making them more in line with base GameObjects (eg. `add.text`).