Fix bug where HP bar could move if hit in the middle of an animation
Whenever a Pokemon takes damage, its HP bar is redrawn.
Part of the `redrawHPBar` function sets the HP bar's location,
because said function is shared with the initialisation.
This resulted in a bug where being hit while in an animation would
move the HP bar (which didn't get moved back).

This commit changes `redrawHPBar` to only redraw the bar (not move it).
The HP bar's position is set in initialisation, and when the Pokemon moves.