Add board organisation (moving Pokemon) during downtime (#7)
This commit adds all the functionality required for moving Pokemon around during downtime.

This is done with a (scene-wide) click handler that uses some new helper functions to
figure out if a click is on a grid tile, and if it contains a Pokemon. Clicks on Pokemon will
select them, and subsequent clicks will move them (if valid).