Fix bug where you couldn't rearrange a full mainboard
There's a check in `movePokemon` which prevents adding more Pokemon
to an already full mainboard, only allowing moving to the mainboard
if it was as part of a swap. The logic prevented moving a Pokemon
from one mainboard square onto another empty one.

This commit fixes that logic so that only sideboard->mainboard
moves can be blocked if the mainboard is full.