Export typed pokemonData instead of raw values
Currently, the `pokemonData` exported is a const, raw object, which TypeScript
knows the exact type of. This means that optional fields in the `Pokemon`
model can't be accessed without a clunky `if ('field' in pokemon)` check,
because otherwise TypeScript will complain that the field doesn't exist on
some of the Pokemon.

This commit changes the exported object in `pokemon.model.ts` to use
the typed pokemonData object. Typescript now knows each entry has the type
`Pokemon`, which means it will allow accessing optional fields.