Refactor turns to be based on Pokemon and not location (#13)
Currently, turns in the CombatScene are taken based on grid coords (x, y)
and not on the actual Pokemon taking the turns. This works fine for now,
but could get unreliable if Pokemon start moving outside of their turns.

This commit refactors the code to take turns for actual PokemonObjects.
The coordinates and stuff are searched for based off the pokemon's id.