Refactor Pokemon stats to make stats consistent across an evolution line (#16)
This commit refactors the Pokemon stats to make stats consistent across an evolution line and to allow for "non-evolving" evolution lines.

To simplify stats and make the mechanics more consistent with traditional autochess games, this commit removes the unique base stats across evolution lines in favour of using the highest evolution's base stats. The stats for an individual Pokemon are calculated based off the base Pokemon and the evolution stage (max HP and attack stats increase for each evolution).

This commit also adds new entries for the full Litwick and Rotom lines.