Support custom targetting logic for moves  (#25)
This commit adds an optional `MoveConfig` field which allows you to provide an optional targeting function for the move. This will be useful for AoE moves which want to maximise targets, or for buffs/healing moves which want to target allies.

If `getFunction` is not provided, it uses the standard `getNearestTarget` as a default.