Add target optimisation for AOE moves (#28)
This commit adds support for optimising AOE moves by adding an `optimiseAOE` function which can be called from a move's custom `getTarget`.

In addition, certain AOE moves also need to be able to target the ground (and not a specific unit) in order to hit the maximum number of targets. This commit also adds differentiation between ground and unit targeted moves, only providing a `target: PokemonObject` to moves which need to target the ground. This is done via parametrising the `ActiveMove` type by `'ground'` or `'unit'`, so the type system enforces that unit targeted moves need a targetPokemon, while ground targeted ones don't. Doing it like this means we don't need `undefined` checks in every move's `use` function,