Save targets between turns (#44)
Currently, targets are calculated at the start of each Pokemon's turn. Pokemon can switch targets if another Pokemon moves closer than their current target. Because the getNearestTarget is a BFS, it can cause a target switch even if the new target is the same distance.

This commit adds a currentTarget field to a PokemonObject which tracks the current target so it can be saved between turns. A few moves which trigger movement have been updated to reset targetting afterwards.