Add Gastly line (#57)
This commit adds the Gastly line, with the move Shadow Ball. I've tried something new in the move logic for Shadow Ball that uses physics collision for damage logic rather than pure-tile-based logic. This was done to make it look nicer (due to the slow projectile) at the potential (??) cost of performance.