Improve synergy rendering performance (#86)
The current synergy display logic destroys/reconstructs them every time they need to be updated. This is quite expensive, as construction requires loading in images. When there are lots of synergies (5-10+), this can cause a few frames to be dropped when moving Pokemon on the board.

This commit updates the logic to render them on-demand and hide them when they're not in use. It makes the necessary refactors to SynergyMarker so that markers can have their visibility toggled.