Add proper fake move with description for Magikarp + remove all changing-type Pokemon
Magikarp's move currently shows as undefined.
This commit fixes it with a custom fake move so it has
both a name and a description.

As an aside, this removes all the Pokemon that change types
(just Magikarp and Scyther).
It was probably a bit confusing and added unnecessary
Pokemon knowledge requirements to the game.