Abstract shop rerolling from ShopScene
Currently, the shop rolling logic is tied to the ShopScene.
When we add AI players, their shops don't need to be rendered,
but they still need to be able to roll.

This commit abstracts the shop reroll function into a pure function,
and sets it up to be called in the ShopScene.