Fix bug where Regigigas crashed the game
Regigigas has 0 base HP because it gets overridden by the HP of the Pivots.
The new HP formula resulted in that creating a negative HP stat.
The game went into an infinite loop trying to render the (negative)
amount of ticks in its HP bar lol

This commit fixes the bug by setting Regigigas' base HP to 100 instead of 0.