Fix bug where Mudkip was treated like a Lapras sometimes
Each Pokemon has a `base` which is used to track what Pokemon
to add back to the shop after selling.
It's also used for AI and synergies.

There was a bug where Mudkip had the wrong Lapras base type,
so selling a mudkip could put a Lapras into the store, and
Mudkip/Lapras would count as duplicates for synergies.

This commit fixes the issue by no longer hardcoding `base`s
and instead generating them as part of the `getEvolution` logic.