Fix a possible error thrown by board prioritisation
If a player has less deduplicated Pokemon than their level,
ie. their only possible boards include duplicates,
the AI logic can throw an "x is undefined" error.

This commit fixes the error by checking if there are no more non-dupes
left, and logs a warn in the console. There's no need to throw or anything,
because this is a perfectly possible/valid thing to happen
eg. a player levels up and their bench is all duplicates