Add generic triggered Effect to be shared by statuses, moves, and synergies (#121)
This commit moves triggered effects out of the Synergy type to a generic Effect type which can be attached to any kind of passive thing. This adds support for all of the synergy triggers to passive moves, and deletes the flag system for passive moves. As a side "effect" (heh), this has made Litwick's move do nothing.