Refactor Moves to improve typings
Previously, I set up Moves so they would be type-validated via `: Move` declarations
so I could use `this.` fields to reference mechanical values.

However, with more optional types like all the Effects on PassiveMoves, the type checking
becomes weaker as it doesn't do TypeScript's innate unkown properties checking when you
declare an object literal. This is bad as you can then declare nonexistent properties
like `onHits` which do nothing.

This commit shifts some of the typings in moves to stop using `this.` so the
objects can be checked to fully and exactly match the Move models.