Add Swinub line + refactor the ppReduction status (#127)
This commit adds the Swinub line (Ice / Ground / Disruptor) with the move Icicle Crash.

It also refactors the "ppReduction" status to be more like mana reave in TFT, where it increases the next move's PP cost by a % instead. This makes it more consistently applicable, as it will always cause the same impact on move use (reducing PP gain has reduced effect if applied when the target has high PP, for example).